+++ pydidas/plugins/base_plugin.py
@@ -4,8 +4,8 @@
 # SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License version 3 as published by
-# the Free Software Foundation.
+# it under the terms of the GNU General Public License version 3 as
+# published by the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -22,8 +22,8 @@ and OutputPlugin base classes.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2021-2022, Malte Storm, Helmholtz-Zentrum Hereon"
-__license__ = "GPL-3.0"
+__copyright__ = "Copyright 2021-, Helmholtz-Zentrum Hereon"
+__license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Development"
 __all__ = ["BasePlugin"]
@@ -272,6 +272,17 @@ class BasePlugin(ObjectWithParameterCollection):
         from .plugin_getter_ import plugin_getter
 
         return (plugin_getter, (self.__class__.__name__,), self.__getstate__())
+
+    def copy(self):
+        """
+        Create a copy of itself.
+
+        Returns
+        -------
+        BasePlugin :
+            The plugins's copy.
+        """
+        return copy.copy(self)
 
     def execute(self, data, **kwargs):
         """